# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: BTP User Containment API
  version: 0.1.0-SNAPSHOT
  description: |
    REST API for the OOBS: a central operator console
    that enrolls SAP BTP subaccounts / IAS tenants, discovers identity providers
    and SoD conflicts, and contains (locks) or unlocks compromised users across subaccounts,
    Cloud Foundry (XSUAA) and IAS in response to an incident.

    ## Authentication
    All `/api/v1/**` endpoints (except `/health` and `/whoami`) require a valid
    XSUAA-issued OAuth2 **bearer JWT**. Authorization is enforced per-operation
    via OAuth2 scopes; the required scope is noted in each operation description.

    ## CSRF
    Keep in mind that the SAP app router uses CSRF. 
    For that see: x-csrf-token docu on the sap help site 
    (it changes all x days so you need to search it)

    ## Scope model
    Four scopes exist, in increasing privilege. The `@PreAuthorize` rules map to:
    - **`btpc.viewer`** - read-only. Satisfied by any of the four scopes.
    - **`btpc.sod_viewer`** - may run SoD scans (admin also allowed).
    - **`btpc.responder`** - may contain/unlock users (admin also allowed).
    - **`btpc.admin`** - full control: enrollment, credentials, config, deletes.

    "Required scope" below names the *minimum* intent; `btpc.admin` always
    satisfies it.

    ## Caller source
    Mutating endpoints accept an optional `X-Caller-Source` header used only for
...
    not change behaviour or authorization.
  license:
...
  contact:
...
servers:
...
security:
  - xsuaa: []
paths:
...
  /api/v1/app-config:
    get:
      tags: [App config]
      summary: List global-account config values
      description: 'Required scope: `btpc.viewer`.'
      responses:
        '200':
          description: Map of config key to its current state
          content:
            application/json:
              schema:
                type: object
                additionalProperties: { $ref: '#/components/schemas/KeyState' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/app-config/{key}:
    put:
      tags: [App config]
      summary: Update a whitelisted config key
      description: |
        Required scope: `btpc.admin`. Only known keys are accepted
        (external-email internal domains, health auto-probe interval); an
        unknown key yields `400`.
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AppConfigUpdate' }
      responses:
        '200':
          description: Updated key state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KeyState' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # -------------------------------------------------------------- Subaccounts
...
  /api/v1/subaccounts/{subaccountId}/contacts:
    parameters:
      - name: subaccountId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Contacts]
      summary: List contacts for a subaccount
      description: 'Required scope: `btpc.viewer`.'
      responses:
        '200':
          description: Contacts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SubaccountContact' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Contacts]
      summary: Add a contact
      description: 'Required scope: `btpc.admin`.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContactPayload' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubaccountContact' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/subaccounts/{subaccountId}/contacts/{id}:
    parameters:
      - name: subaccountId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - $ref: '#/components/parameters/IdPath'
    put:
      tags: [Contacts]
      summary: Update a contact
      description: 'Required scope: `btpc.admin`.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContactPayload' }
      responses:
        '200':
          description: Updated contact
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubaccountContact' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Contacts]
      summary: Delete a contact
      description: 'Required scope: `btpc.admin`.'
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # --------------------------------------------------------------- IAS tenants
...
  /api/v1/containment:
    post:
      tags: [Containment]
      summary: Contain (lock) a user
      description: |
        Required scope: `btpc.responder`. Locks a user across the selected
        subaccounts / IAS tenants per the action flags in the request.
      parameters:
        - $ref: '#/components/parameters/CallerSource'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContainRequest' }
      responses:
        '200':
          description: Containment result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContainmentResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '502': { $ref: '#/components/responses/BadGateway' }
  /api/v1/containment/unlock-preview:
    get:
      tags: [Containment]
      summary: Preview what an unlock would restore
      description: 'Required scope: `btpc.viewer`.'
      parameters:
        - name: userEmail
          in: query
          required: true
          schema: { type: string }
        - name: subaccountIds
          in: query
          required: false
          schema:
            type: array
            items: { type: string, format: uuid }
          style: form
          explode: true
      responses:
        '200':
          description: Preview entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/UnlockPreviewEntry' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/containment/unlock:
    post:
      tags: [Containment]
      summary: Unlock a previously contained user
      description: 'Required scope: `btpc.responder`.'
      parameters:
        - $ref: '#/components/parameters/CallerSource'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UnlockRequest' }
      responses:
        '200':
          description: Unlock result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UnlockResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '502': { $ref: '#/components/responses/BadGateway' }

  # ----------------------------------------------------------- Protected users
...
  /api/v1/audit-sinks:
    get:
      tags: [Audit sinks]
      summary: List configured audit sinks
      description: 'Required scope: `btpc.viewer`. Secrets are never returned (only `hasSecret`).'
      responses:
        '200':
          description: Audit sinks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AuditSinkDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/audit-sinks/{kind}:
    parameters:
      - name: kind
        in: path
        required: true
        description: Sink kind (e.g. splunk_hec, webhook)
        schema: { type: string }
    put:
      tags: [Audit sinks]
      summary: Create or update an audit sink
      description: 'Required scope: `btpc.admin`.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuditSinkUpsert' }
      responses:
        '200':
          description: Saved sink
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuditSinkDto' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: [Audit sinks]
      summary: Delete an audit sink
      description: 'Required scope: `btpc.admin`.'
      responses:
        '200': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/audit-sinks/{kind}/test:
    parameters:
      - name: kind
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Audit sinks]
      summary: Test an audit sink connection
      description: 'Required scope: `btpc.admin`. Always returns 200; the `ok` flag reports success.'
      responses:
        '200':
          description: Test result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuditSinkTestResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    xsuaa:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        XSUAA-issued OAuth2 access token. Scopes (`btpc.admin`,
        `btpc.responder`, `btpc.sod_viewer`, `btpc.viewer`) are carried as
        `SCOPE_<name>` authorities and enforced per operation.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    CallerSource:
      name: X-Caller-Source
      in: header
      required: false
      description: Audit attribution only; default UI.
      schema:
        type: string
        enum: [soar, soar-api, system]
  responses:
    BadRequest:
      description: Invalid input or malformed JSON body
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid bearer token
    Forbidden:
      description: Authenticated but missing the required scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: State conflict or data-integrity violation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BadGateway:
      description: Upstream BTP/CIS failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      description: Uniform error envelope. `null` fields are omitted.
      properties:
        error:
          type: string
          description: Stable machine code.
          enum: [bad_request, malformed_body, not_found, conflict, cis_upstream, forbidden, server_error]
        message:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time

    # ---- Platform ----
...
    AppConfigUpdate:
      type: object
      required: [value]
      properties:
        value: { type: string }
    KeyState:
      type: object
      properties:
        value: { type: string }
        overridden:
          type: boolean
          description: True when an explicit override exists (vs. the default).

    # ---- Subaccount request bodies ----
...
    ContactPayload:
      type: object
      description: ContactService.ContactPayload.
      required: [name, email, role]
      properties:
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [security, ops, business, technical, other] }
        notes: { type: string, nullable: true }
    SubaccountContact:
      type: object
      description: ninja.gruber.btpc.domain.SubaccountContact.
      properties:
        id: { type: string, format: uuid }
        subaccountId: { type: string, format: uuid }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [security, ops, business, technical, other] }
        notes: { type: string, nullable: true }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time, nullable: true }

    # ---- IAS tenants ----
...
    ContainRequest:
      type: object
      description: ContainmentDtos.ContainRequest.
      required: [actions, userEmail]
      properties:
        actions:
          type: array
          minItems: 1
          description: >
            Required, non-empty. Per-subaccount actions: `xsuaa_strip_roles`,
            `xsuaa_delete_shadow`, `cf_revoke_org_roles`. Per-IAS-tenant actions:
            `ias_deactivate`, `ias_strip_groups`.
          items:
            type: string
            enum: [xsuaa_strip_roles, xsuaa_delete_shadow, cf_revoke_org_roles, ias_deactivate, ias_strip_groups]
        userEmail: { type: string }
        subaccountIds:
          type: array
          nullable: true
          description: Null/empty → all enrolled subaccounts (only used if a per-SA action is requested).
          items: { type: string, format: uuid }
        iasTenantIds:
          type: array
          nullable: true
          description: Null/empty → all enrolled IAS tenants (only used if an IAS action is requested).
          items: { type: string, format: uuid }
        originMode: { type: string, enum: [all, list, discovered], description: How to resolve which XSUAA origins to act on. }
        origins:
          type: array
          nullable: true
          description: Explicit origin keys; used when originMode = list.
          items: { type: string }
        dryRun: { type: boolean, default: true, description: Defaults to true - plan only, no mutation.}
        comment: { type: string, nullable: true }
    ActionResult:
      type: object
      description: Outcome of one containment/unlock action against one system.
      properties:
        action: { type: string }
        outcome: { type: string, description: 'Lower-kebab outcome, e.g. ok, failed, skipped, dry-run.' }
        message: { type: string, nullable: true }
        snapshotId:
          type: string
          format: uuid
          nullable: true
          description: Snapshot captured for later unlock/restore, when applicable.
    SubaccountActionResult:
      type: object
      description: Per-subaccount slice of a containment/unlock result.
      properties:
        subaccountId: { type: string, format: uuid }
        displayName: { type: string }
        blockedByProtection: { type: boolean }
        results:
          type: array
          items: { $ref: '#/components/schemas/ActionResult' }
        protectionReasons:
          type: array
          items: { type: string }
    ContainmentResult:
      type: object
      description: ContainmentDtos.ContainmentResult.
      properties:
        correlationId: { type: string, format: uuid }
        dryRun: { type: boolean }
        resolvedIasTenantIds:
          type: array
          items: { type: string, format: uuid }
        iasResults:
          type: array
          description: One entry per selected IAS tenant.
          items: { $ref: '#/components/schemas/ActionResult' }
        perSubaccount:
          type: array
          items: { $ref: '#/components/schemas/SubaccountActionResult' }
        globalProtectionReasons:
          type: array
          nullable: true
          description: >
            Non-empty → the request was hard-blocked by a global protection;
            iasResults and perSubaccount are then empty.
          items: { type: string }
    UnlockRequest:
      type: object
      description: ContainmentDtos.UnlockRequest.
      required: [actions, userEmail]
      properties:
        actions:
          type: array
          minItems: 1
          description: >
            Required, non-empty. Per-subaccount actions: `xsuaa_restore_roles`,
            `cf_restore_org_roles`. Per-IAS-tenant actions: `ias_activate`,
            `ias_restore_groups`.
          items:
            type: string
            enum: [xsuaa_restore_roles, cf_restore_org_roles, ias_activate, ias_restore_groups]
        userEmail: { type: string }
        subaccountIds:
          type: array
          nullable: true
          items: { type: string, format: uuid }
        iasTenantIds:
          type: array
          nullable: true
          items: { type: string, format: uuid }
        correlationId:
          type: string
          format: uuid
          nullable: true
          description: Containment run to restore from; null restores the latest snapshot per target.
        dryRun: { type: boolean, default: true }
        comment: { type: string, nullable: true }
    UnlockResult:
      type: object
      description: ContainmentDtos.UnlockResult.
      properties:
        correlationId: { type: string, format: uuid }
        dryRun: { type: boolean }
        resolvedIasTenantIds:
          type: array
          items: { type: string, format: uuid }
        iasResults:
          type: array
          items: { $ref: '#/components/schemas/ActionResult' }
        perSubaccount:
          type: array
          items: { $ref: '#/components/schemas/SubaccountActionResult' }
    UnlockPreviewEntry:
      type: object
      description: ContainmentDtos.UnlockPreviewEntry - what a restore would re-grant in one subaccount.
      properties:
        subaccountId: { type: string, format: uuid }
        displayName: { type: string }
        snapshotId: { type: string, format: uuid }
        perOrigin:
          type: array
          items: { $ref: '#/components/schemas/UnlockPreviewOrigin' }
        totalRoles: { type: integer }
        note: { type: string, nullable: true }
    UnlockPreviewOrigin:
      type: object
      properties:
        origin: { type: string }
        roleCollections:
          type: array
          items: { type: string }

    # ---- Protected users ----
...
    AuditSinkUpsert:
      type: object
      properties:
        enabled: { type: boolean }
        config:
          type: object
          description: Sink-specific config (JSON).
          additionalProperties: true
        secret:
          type: string
          nullable: true
          description: Write-only secret; never returned on read.
    AuditSinkDto:
      type: object
      properties:
        kind: { type: string }
        enabled: { type: boolean }
        config:
          type: object
          additionalProperties: true
        hasSecret: { type: boolean }
        lastTestAt: { type: string, format: date-time, nullable: true }
        lastTestStatus: { type: string, nullable: true }
        lastTestMessage: { type: string, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
        updatedBy: { type: string, nullable: true }
    AuditSinkTestResult:
      type: object
      properties:
        ok: { type: boolean }
        message: { type: string, nullable: true }